+++ src/sublimekeys/__init__.py
@@ -1,8 +1,8 @@
 from .client import LicenseResult, SublimeKeysClient, TrialResult
-from .exceptions import LeaseError, NetworkError, SublimeKeysError
+from .exceptions import LeaseError, NetworkError, ServerError, SublimeKeysError
 from .machine import get_or_create_machine_id
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 __all__ = [
     "SublimeKeysClient",
@@ -10,6 +10,7 @@ __all__ = [
     "TrialResult",
     "SublimeKeysError",
     "NetworkError",
+    "ServerError",
     "LeaseError",
     "get_or_create_machine_id",
 ]